requirements.txt
setuptools
python-dateutil
pytz
nose
numpy
scipy
xarray
h5py
astropy
matplotlib
seaborn
